meridianoTickets/requirements.txt
Django=2.1
pyodbc
django-mssql-backend
djangorestframework
markdown
